pyproject.toml
[tool.poetry]
name = "lms"
version = "0.1.0"
description = "lms"
authors = ["SergeiNaum <sergei.naumencko@yandex.ru>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
gunicorn = "^21.2.0"
drf-spectacular = "^0.27.1"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
black = "^24.1.1"
ipython = "^8.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
